Replace `EncodePacket` and `DecodePacket` with `Packet`
