+++ crates/packet_inspector/src/main.rs
@@ -15,6 +15,7 @@ use tokio::net::{TcpListener, TcpStream};
 use tokio::sync::Semaphore;
 use tokio::task::JoinHandle;
 use tracing_subscriber::filter::LevelFilter;
+use valence_protocol::Packet;
 use valence_protocol::codec::{PacketDecoder, PacketEncoder};
 use valence_protocol::packet::c2s::handshake::handshake::NextState;
 use valence_protocol::packet::c2s::handshake::HandshakeC2s;
@@ -23,7 +24,6 @@ use valence_protocol::packet::c2s::status::{QueryPingC2s, QueryRequestC2s};
 use valence_protocol::packet::s2c::login::LoginSuccessS2c;
 use valence_protocol::packet::s2c::status::{QueryPongS2c, QueryResponseS2c};
 use valence_protocol::packet::{C2sPlayPacket, S2cLoginPacket, S2cPlayPacket};
-use valence_protocol::{DecodePacket, EncodePacket};
 
 #[derive(Parser, Clone, Debug)]
 #[clap(author, version, about)]
@@ -65,7 +65,7 @@ struct State {
 impl State {
     pub async fn rw_packet<'a, P>(&'a mut self) -> anyhow::Result<P>
     where
-        P: DecodePacket<'a> + EncodePacket,
+        P: Packet<'a>,
     {
         while !self.dec.has_next_packet()? {
             self.dec.reserve(4096);
